refactor: move intro map to maps
