+++ src/maps.ts
@@ -83,3 +83,37 @@ export const createMap = (number: number): Array2D<Tile> => {
 
     return grid;
 };
+
+export const createIntroMap = (): Array2D<Tile> => {
+    const introWidth = 10;
+    const introHeight = 18;
+    const grid = new Array2D<Tile>(introWidth, introHeight);
+    const centerCol = Math.floor(introWidth / 2);
+    for (let iy = 0; iy < introHeight; iy++) {
+        for (let ix = 0; ix < introWidth; ix++) {
+            const x = ix * TILE_SIZE;
+            const y = iy * TILE_DRAW_HEIGHT;
+            let type = TileType.Grass;
+            // Place bushes and flowers at fixed positions, not in the center column
+            if (
+                (iy === 2 && ix === 2) ||
+                (iy === 2 && ix === introWidth - 3) ||
+                (iy === 4 && ix === 1) ||
+                (iy === 4 && ix === introWidth - 2)
+            ) {
+                type = TileType.Bush;
+            } else if (
+                (iy === 3 && ix === 1) ||
+                (iy === 3 && ix === introWidth - 2) ||
+                (iy === 1 && ix === 3) ||
+                (iy === 1 && ix === introWidth - 4)
+            ) {
+                type = TileType.Flower;
+            }
+            // Never place anything in the center column
+            if (ix === centerCol) type = TileType.Slate;
+            grid.setValue(ix, iy, createTile(type, x, y));
+        }
+    }
+    return grid;
+};
